working...

+++ juicefs.sh
@@ -15,9 +15,11 @@ echo "///////////////////////////////////////////////////////////////////////"
 export JUICE_TOKEN=${JUICE_TOKEN:-XXXXX}
 
 # install juicefs
-wget -q https://juicefs.com/static/juicefs
-chmod +x juicefs 
-sudo mv juicefs /usr/bin
+if [[ ! -f /usr/bin/juicefs ]] ; then
+    wget -q https://juicefs.com/static/juicefs
+    chmod +x juicefs 
+    sudo mv juicefs /usr/bin
+fi
 
 # create the bucket if necessary
 aws s3 mb s3://${BUCKET_NAME} --region ${AWS_DEFAULT_REGION}
